kivy-requirements.txt
kivy-garden.qrcode
opencv-python
pyzbar
telenium